(#11628) Fix for crash when other mods create terminals (#194)
For mod interactions with AE2 (namely, Thaumic Energistics (ThE) ) the back button crashes since the OriginalGui is not explicitly listed in the constructor. This adds a null check so the implementing mod can handle the behavior instead.

This commit also adds an accessor for the cancel button so child mods can capture its behavior. This is not a functional change to AE2, but required to fix the crash in ThE and any other child mods

Tested and verified in local server.