appeng/core/sync/packets: Add PacketIfaceTermUpdate.java
Adds a new packet specifically for updating the entries on the terminal.
Reduces network traffic, potentially significantly, as it allows for
incremental updates instead of sending a fresh list every time the entries
on the list change (Note it WAS smart enough to not send a fresh list if
only patterns changed).

Deprecates appeng/core/sync/packets/PacketCompressedNBT

Commit does not add functionality to handling on the client side yet.